ajout des d2i + gestion de sécurités déclenchées par les modérateurs + manipulation des sadidas + correction d'un bug lié aux Int64 + de nombreuses autres fonctionnalités

+++ messages/fights/GameActionFightCastRequestMessage.java
@@ -12,7 +12,7 @@ public class GameActionFightCastRequestMessage extends Message {
 		super();
 	}
 
-	public void serialize(int spellId, short cellId, int instanceId) {
+	public void serialize(int spellId, int cellId, int instanceId) {
 		ByteArray buffer = new ByteArray();
 		buffer.writeVarShort(spellId);
 		buffer.writeShort(cellId);
